add quota project attribute back that was inadvertently removed during a sync with main, fix more circular imports
